==> notebooks/1_preprocessing.ipynb <==
import pandas as pd
from sklearn.feature_extraction import DictVectorizer as DV
import seaborn as sns

%pylab inline

abalone = pd.read_csv('abalone.data', 
                   names= ['Sex','Length','Diameter','Height','Whole weight','Shucked weight','Viscera weight','Shell weight','Rings'])
abalone.head()

encoder = DV(sparse = False)
encoded_sign_abalone = encoder.fit_transform(abalone[['Sex']].T.to_dict().values())
print(encoded_sign_abalone[:10])

abalone = abalone.drop('Sex',1)
abalone_encoded = pd.concat([pd.DataFrame(encoded_sign_abalone,columns=['Female','Infant','Male']),abalone],axis = 1)
abalone_encoded.head()

abalone_encoded.to_csv('abalone_preprocessed.csv')

concrete = pd.read_excel('concrete.xls', names = ['Cement', 'Blast Furnace Slag', 'Fly Ash', 'Water','Superplasticizer',
                                                  'Coarse Aggregate', 'Fine Aggregate', 'Age', 'Concrete compressive strength'])
concrete.head()

concrete.columns

concrete.to_csv('concrete_preprocessed.csv')
