==> notebooks/2_classification.ipynb <==
import pandas as pd
from sklearn.model_selection import train_test_split
from sklearn.preprocessing import StandardScaler
from sklearn.neighbors import KNeighborsClassifier
from sklearn.metrics import classification_report, confusion_matrix
from sklearn import tree
from sklearn.naive_bayes import GaussianNB
from sklearn import svm
from sklearn.linear_model import LogisticRegression 
from sklearn.model_selection import GridSearchCV

abalone = pd.read_csv('../data/abalone_preprocessed.csv', index_col=0)
abalone.columns


abalone.head()

abalone["Rings"].unique()

for i in range(1,30):
  print(abalone[abalone.Rings == i].shape[0])

abalone.info()

Разобьём выборку на категориальные признаки, числовые признаки и на классы.

X_cat = abalone[['Female','Infant', 'Male']]
y = pd.DataFrame(abalone.iloc[:,-1])
X_real = abalone.iloc[:, 3:-1]

y.head()

Заменим значения y на возрастной диапазон к которому относится экзэмпляр. С 1 до 7 колец - молодой, с 8 до 11 средний, с 12 до 29 старый.

for i in range(len(y)):
  if(y['Rings'][i] < 8):
    y['Rings'][i] = "young"
  elif(y['Rings'][i] < 12):
    y['Rings'][i] = "middle"
  else:
    y['Rings'][i] = "old"

y.head()

(X_real_train, X_real_test, y_train, y_test) =  train_test_split(X_real, y, test_size = 0.2, random_state = 0)

(X_cat_train, X_cat_test) = train_test_split(X_cat, test_size=0.2, random_state = 0)

scaler = StandardScaler()
scaler.fit(X_real_train, y_train)
X_real_train_scaled = scaler.transform(X_real_train)
X_real_test_scaled = scaler.transform(X_real_test)

X_real_train_scaled

X_cat_train = pd.DataFrame(X_cat_train)
X_cat_test = pd.DataFrame(X_cat_test)

X_real_train_scaled = pd.DataFrame(X_real_train_scaled)
X_real_test_scaled = pd.DataFrame(X_real_test_scaled)

X_cat_train.index = X_real_train_scaled.index
X_train = pd.concat([X_real_train_scaled, X_cat_train], axis=1)

X_cat_test.index = X_real_test_scaled.index
X_test = pd.concat([X_real_test_scaled, X_cat_test], axis=1)

X_train.head()

Обучим модель метода k-ближайших соседей.

knn = KNeighborsClassifier(n_neighbors=5)
knn.fit(X_train,y_train)
y_pred = knn.predict(X_test)
print(confusion_matrix(y_test, y_pred))
print(classification_report(y_test, y_pred))


Попробуем подобрать параметры.

knn.get_params().keys()

knn_parameters_grid = {'n_neighbors': range(1, 10),
                       'weights': ['uniform', 'distance'],
                       'metric': ['minkowski', 'manhattan', 'euclidean', 'chebyshev'],
                       'algorithm': ['brute', 'ball_tree', 'KD_tree', 'auto']
                       }

knn_grid = GridSearchCV(knn, knn_parameters_grid,
                        cv=5, n_jobs=-1, verbose=True)
knn_grid.fit(X_train,y_train)
knn_grid.best_params_, knn_grid.best_score_

Удалось немного повысить точность предсказания.

Теперь обучим классификатор дерева решений.

dtc = tree.DecisionTreeClassifier()
dtc.fit(X_train,y_train)
y_pred_2 = dtc.predict(X_test)
print(confusion_matrix(y_test, y_pred_2))
print(classification_report(y_test, y_pred_2))

Попробуем подобрать гиперпараметры для этой модели.

dtc.get_params().keys()

dtc_parameters_grid = {'max_depth': range(1,6),
                      'max_features': range(1,7),
                      'min_samples_leaf' : range(1,10)}

dtc_grid = GridSearchCV(dtc, dtc_parameters_grid,
                        cv=5, n_jobs=-1, verbose=True)
dtc_grid.fit(X_train, y_train)
dtc_grid.best_params_, dtc_grid.best_score_

Точность увеличилась сильнее чем у прошлой модели после подбора параметров, но всё же оказалась менее точной.

Попробуем сделать то же с наивным байесовским классификатором

nb = GaussianNB()
nb.fit(X_train, y_train)
y_pred_3 = nb.predict(X_test)
print(confusion_matrix(y_test, y_pred_3))
print(classification_report(y_test, y_pred_3))

Точность получилась маленькая.

Проверим метод опорных векторов.

svm = svm.SVC()
svm.fit(X_train, y_train)
y_pred_4 = svm.predict(X_test)
print(confusion_matrix(y_test, y_pred_4))
print(classification_report(y_test, y_pred_4))

Точность изначально выше чем у моделей, стоящих раньше.

svm.get_params().keys()

svm_parameters_grid = {'C': [0.1, 1, 10, 100, 1000], 

              'gamma': [1, 0.1, 0.01, 0.001, 0.0001],

              'kernel': ['rbf']}

svm_grid = GridSearchCV(svm, svm_parameters_grid,
                        cv=5, n_jobs=-1, verbose=True)
svm_grid.fit(X_train, y_train)
svm_grid.best_params_, svm_grid.best_score_

После подбора гиперпараметров эта модель показала самую высокую точность.

Проверим последнюю модель.

lg = LogisticRegression()
lg.fit(X_train, y_train)
y_pred_5 = lg.predict(X_test)
print(confusion_matrix(y_test, y_pred_5))
print(classification_report(y_test, y_pred_5))

lg.get_params().keys()

lg_parameters_grid = {'penalty' : ['l1', 'l2'],
                       'C' : [1.0, 10, 100, 200],
                       'max_iter': [100,  200, 300],
                       }

lg_grid = GridSearchCV(lg, lg_parameters_grid,
                        cv=5, n_jobs=-1, verbose=True)
lg_grid.fit(X_train, y_train)
lg_grid.best_params_, lg_grid.best_score_

Результат немного меньше чем у прошлой модели, но выше чем у остальных.
