==> notebooks/3_regression.ipynb <==
import pandas as pd
import numpy as np
from sklearn.model_selection import train_test_split
from sklearn.linear_model import LinearRegression
from sklearn.metrics import r2_score, mean_squared_error, mean_absolute_error
from sklearn.preprocessing import StandardScaler
from sklearn.tree import DecisionTreeRegressor
from sklearn.model_selection import GridSearchCV
from sklearn.linear_model import Lasso
from sklearn.linear_model import Ridge
from sklearn.linear_model import ElasticNet

concrete_frame = pd.read_csv('../data/concrete_preprocessed.csv',index_col=0)
concrete_frame.head()

concrete_frame.info()

x = concrete_frame.iloc[:,:-1]
y = concrete_frame.iloc[:,-1]

x.head()

y.head()

(x_train, x_test, y_train, y_test) = train_test_split(x, y,test_size = 0.2, random_state = 0)

scaler = StandardScaler()
scaler.fit(x_train, y_train)
x_train_scaled = scaler.transform(x_train)
x_test_scaled = scaler.transform(x_test)

## Линейная регрессия

Осуществляется с помощью sklearn.linear_model.LinearRegression.

Ссылка на документацию: https://scikit-learn.org/stable/modules/generated/sklearn.linear_model.LinearRegression.html

Обучаем регрессор LineatRegression

lr = LinearRegression(normalize=True,n_jobs=-1)
lr.fit(x_train, y_train)
y_pred_lr = lr.predict(x_test)
print(r2_score(y_test, y_pred_lr))
print(mean_squared_error(y_test, y_pred_lr))
print(mean_squared_error(y_test, y_pred_lr, squared=False))
print(mean_absolute_error(y_test, y_pred_lr))

Модель показывает не хороший результат. Так же нельзя подобрать гиперпараметры чтобы улучшить её.

## Регрессия дерева решений

Осуществляется с помощью sklearn.tree.DecisionTreeRegressor.

Ссылка на документацию: https://scikit-learn.org/stable/modules/generated/sklearn.tree.DecisionTreeRegressor.html



dtr = DecisionTreeRegressor()
dtr.fit(x_train, y_train)
y_pred_dtr = dtr.predict(x_test)
print(r2_score(y_test, y_pred_dtr))
print(mean_squared_error(y_test, y_pred_dtr))
print(mean_squared_error(y_test, y_pred_dtr, squared=False))
print(mean_absolute_error(y_test, y_pred_dtr))

Дерево решений показало результаты лучше, чем линейнаяы регрессия. Попробуем подобрать гиперпараметры, чтобы улучшить результат.

Воспользуемся sklearn.model_selection.GridSearchCV

dtr.get_params().keys()

dtr_parameters_grid = {'criterion' : ['mse', 'friedman_mse', 'mae', 'poisson'],
                       'splitter' : ['best', 'random'],
                       'max_depth' : range(1, 11),
                       'min_samples_split' : range(1, 10) 
}

dtr_grid = GridSearchCV(dtr, dtr_parameters_grid,
                        cv=5, n_jobs=-1, verbose=True)
dtr_grid.fit(x_train, y_train)
dtr_grid.best_params_

dtr1 = DecisionTreeRegressor(criterion='mse', max_depth=10, splitter='best',min_samples_split=4)
dtr1.fit(x_train, y_train)
y_pred_dtr1 = dtr1.predict(x_test)
print(r2_score(y_test, y_pred_dtr1))
print(mean_squared_error(y_test, y_pred_dtr1))
print(mean_squared_error(y_test, y_pred_dtr1, squared=False))
print(mean_absolute_error(y_test, y_pred_dtr1))

Результаты улучшились, но не значительно.


## LASSO регрессия

Осуществляется с помощью sklearn.linear_model.Lasso.

Ссылка на документацию: https://scikit-learn.org/stable/modules/generated/sklearn.linear_model.Lasso.html

lasso = Lasso()
lasso.fit(x_train,y_train)
y_pred_lasso = lasso.predict(x_test)
print(r2_score(y_test, y_pred_lasso))
print(mean_squared_error(y_test, y_pred_lasso))
print(mean_squared_error(y_test, y_pred_lasso, squared=False))
print(mean_absolute_error(y_test, y_pred_lasso))

Результаты схожи с линейной регрессией. Попробуем подобрать параметры с помощью GridSearchCV

lasso.get_params().keys()

lasso_parameters_grid = {'alpha' : np.arange(0.01, 1.0, 0.05),
                         'normalize' : [True, False],
                         'max_iter' : [10, 50 ,100, 200, 400, 800, 1000, 1500, 2000],
                         'selection' : ['cyclic', 'random'],
                         'warm_start':[True,False]
                         }

lasso_grid = GridSearchCV(lasso, lasso_parameters_grid, cv = 5, n_jobs=-1, verbose=True)
lasso_grid.fit(x_train, y_train)
lasso_grid.best_params_

lasso1 = Lasso(alpha=0.11, max_iter=100, selection='random')
lasso1.fit(x_train,y_train)
y_pred_lasso1 = lasso.predict(x_test)
print(r2_score(y_test, y_pred_lasso1))
print(mean_squared_error(y_test, y_pred_lasso1))
print(mean_squared_error(y_test, y_pred_lasso1, squared=False))
print(mean_absolute_error(y_test, y_pred_lasso1))

После подбора параметров результат никак не изменился.

## Гребневая регрессия

Осуществляется с помощью sklearn.linear_model.Ridge.

Ссылка на документацию: https://scikit-learn.org/stable/modules/generated/sklearn.linear_model.Ridge.html

---



ridge = Ridge()
ridge.fit(x_train_scaled,y_train)
y_pred_ridge = ridge.predict(x_test_scaled)
print(r2_score(y_test, y_pred_ridge))
print(mean_squared_error(y_test, y_pred_ridge))
print(mean_squared_error(y_test, y_pred_ridge, squared=False))
print(mean_absolute_error(y_test, y_pred_ridge))

Результат схож с линейной и lasso регрессией. Подберём параметры по сетке.

ridge.get_params().keys()

ridge_parameters_grid = {'alpha' : np.arange(0.01, 1.0, 0.05),
                         'max_iter' : [10, 50 ,100, 200, 400, 800, 1000, 1500, 2000],
                         'solver' : ['auto', 'svd', 'cholesky', 'lsqr', 'sparse_cg', 'sag', 'saga'],
                         'fit_intercept' : [True, False]
                         }

ridge_grid = GridSearchCV(ridge, ridge_parameters_grid, cv = 5, n_jobs=-1, verbose=True)
ridge_grid.fit(x_train_scaled, y_train)
ridge_grid.best_params_

ridge1 = Ridge(alpha=0.21, max_iter=10,solver='sag')
ridge1.fit(x_train_scaled,y_train)
y_pred_ridge1 = ridge1.predict(x_test_scaled)
print(r2_score(y_test, y_pred_ridge1))
print(mean_squared_error(y_test, y_pred_ridge1))
print(mean_squared_error(y_test, y_pred_ridge1, squared=False))
print(mean_absolute_error(y_test, y_pred_ridge1))

Результат незначительно вырос.

## Elastic Net регрессия

Осуществляется с помощью sklearn.linear_model.ElasticNet

Ссылка на документацию: https://scikit-learn.org/stable/modules/generated/sklearn.linear_model.ElasticNet.html

en = ElasticNet()
en.fit(x_train, y_train)
y_pred_en = en.predict(x_test)
print(r2_score(y_test, y_pred_en))
print(mean_squared_error(y_test, y_pred_en))
print(mean_squared_error(y_test, y_pred_en, squared=False))
print(mean_absolute_error(y_test, y_pred_en))

Elastic Net регрессия выдаёт не лучший результат. Попробуем улучшить, подбирая параметры.

en.get_params().keys()

en_parameters_grid = {'alpha' : np.arange(0.01, 1.0, 0.05),
                         'l1_ratio' : np.arange(0.1, 1.0, 0.1),
                         'max_iter' : [10, 50 ,100, 200, 400, 800, 1000],
                         'fit_intercept' : [True, False],
                         'normalize' : [True, False],
                         'selection' : ['cyclic', 'random'],
                         }

en_grid = GridSearchCV(en, en_parameters_grid, cv = 5, n_jobs=-1, verbose=True)
en_grid.fit(x_train, y_train)
en_grid.best_params_

en1 = ElasticNet(alpha=0.51, fit_intercept=False, max_iter=200, normalize=True,selection='random')
en1.fit(x_train, y_train)
y_pred_en1 = en1.predict(x_test)
print(r2_score(y_test, y_pred_en1))
print(mean_squared_error(y_test, y_pred_en1))
print(mean_squared_error(y_test, y_pred_en1, squared=False))
print(mean_absolute_error(y_test, y_pred_en1))

С подобранными параметрами результат немного улучшился, но подбор был слишком долгим.

Из всех алгоритмов регрессии, регрессия дерева решений лучше всего показала себя на данном датасете.
