==> notebooks/4_ensemble.ipynb <==
import pandas as pd
import numpy as np
from sklearn.svm import SVC
from sklearn.tree import DecisionTreeRegressor
from sklearn.model_selection import train_test_split
from sklearn.preprocessing import StandardScaler
from sklearn.model_selection import GridSearchCV
from sklearn import metrics
from sklearn import ensemble  
from sklearn.neighbors import KNeighborsClassifier
from sklearn.linear_model import Ridge

# Композиции модели для задачи классификации

## Подготовка данных

abalone = pd.read_csv('../data/abalone_preprocessed.csv', index_col=0)
abalone.head()

X_cat = abalone[['Female','Infant', 'Male']]
y = pd.DataFrame(abalone.iloc[:,-1])
X_real = abalone.iloc[:, 3:-1]

for i in range(len(y)):
  if(y['Rings'][i] < 8):
    y['Rings'][i] = "young"
  elif(y['Rings'][i] < 12):
    y['Rings'][i] = "middle"
  else:
    y['Rings'][i] = "old"

(X_real_train, X_real_test, y_train, y_test) =  train_test_split(X_real, y, test_size = 0.2, random_state = 0)

(X_cat_train, X_cat_test) = train_test_split(X_cat, test_size=0.2, random_state = 0)

scaler = StandardScaler()
scaler.fit(X_real_train, y_train)
X_real_train_scaled = scaler.transform(X_real_train)
X_real_test_scaled = scaler.transform(X_real_test)

X_cat_train = pd.DataFrame(X_cat_train)
X_cat_test = pd.DataFrame(X_cat_test)

X_real_train_scaled = pd.DataFrame(X_real_train_scaled)
X_real_test_scaled = pd.DataFrame(X_real_test_scaled)

X_cat_train.index = X_real_train_scaled.index
X_train = pd.concat([X_real_train_scaled, X_cat_train], axis=1)

X_cat_test.index = X_real_test_scaled.index
X_test = pd.concat([X_real_test_scaled, X_cat_test], axis=1)

## Обучение модели

В качестве базового алгоритма используем метод опорных векторов, потому что он лучше всего показал себя на этой выборке. Используем лучшие параметры полученные в лабораторной работе № 2.



svm = SVC(C = 1000, gamma = 0.01)
svm.fit(X_train, y_train)
y_pred_svm = svm.predict(X_test)
print(metrics.confusion_matrix(y_test, y_pred_svm))
print(metrics.classification_report(y_test, y_pred_svm))

## Обучение композиций

### Бэггинг

%%time
bagging = ensemble.BaggingClassifier(svm,n_jobs = -1)
bagging.fit(X_train, y_train)
pred_bagging = bagging.predict(X_test)
print(metrics.confusion_matrix(y_test, pred_bagging))
print(metrics.classification_report(y_test, pred_bagging))

Подберём параметры бэггинга по сетке с помощью GridSearchCV.

bagging_parameters_grid = { 'n_estimators' : [1, 5, 10, 50, 100],
                           'max_features' : range(1,10,1)
    
}

bagging_grid = GridSearchCV(bagging, bagging_parameters_grid,
                        cv=5, n_jobs=-1, verbose=True)
bagging_grid.fit(X_train, y_train)
bagging_grid.best_params_, bagging_grid.best_score_

Точность немного улучшилась, но на поиск параметров ушло очень много времени.

### Бустинг

%%time
boosting = ensemble.GradientBoostingClassifier()
boosting.fit(X_train, y_train)
pred_boosting = boosting.predict(X_test)
print(metrics.confusion_matrix(y_test, pred_boosting))
print(metrics.classification_report(y_test, pred_boosting))

boosting_parameters_grid = { 'learning_rate' : np.arange(0.1, 0.8, 0.1),
                           'n_estimators' : [10, 50, 100],
                            'max_depth' : range(1,10,1)

}

boosting_grid = GridSearchCV(boosting, boosting_parameters_grid,
                        cv=5, n_jobs=-1, verbose=True)
boosting_grid.fit(X_train, y_train)
boosting_grid.best_params_, boosting_grid.best_score_

Результат не изменился

### Стекинг

estimators = [
('svm', SVC()),
('knn', KNeighborsClassifier(algorithm = 'brute',metric = 'manhattan', n_neighbors = 9))
]

%%time
stacking = ensemble.StackingClassifier(estimators=estimators)
stacking.fit(X_train, y_train)
pred_stacking = stacking.predict(X_test)
print(metrics.confusion_matrix(y_test, pred_stacking))
print(metrics.classification_report(y_test, pred_stacking))

stacking_parameters_grid = { 'stack_method' : ['auto', 'pred_proba', 'solution_function', 'predict'],
                            'cv' : [1, 3, 5, 7, 9]
    
}

stacking_grid = GridSearchCV(stacking, stacking_parameters_grid,
                        cv=5, n_jobs=-1, verbose=True)
stacking_grid.fit(X_train, y_train)
stacking_grid.best_params_, stacking_grid.best_score_

После подбора параметров композиция оказалась немного хуже остальных, но быстрее. 

Лучшей композицией оказался бэггинг.

# Композиции модели для задачи регрессии

## Подготовка данных

concrete = pd.read_csv('../data/concrete_preprocessed.csv',index_col=0)
concrete.head()

X_reg = concrete.iloc[:,:-1]
y_reg = concrete.iloc[:,-1]

(X_reg_train, X_reg_test, y_reg_train, y_reg_test) = train_test_split(X_reg, y_reg,test_size = 0.2, random_state = 0)

## Обучение модели

dtr = DecisionTreeRegressor()
dtr.fit(X_reg_train, y_reg_train)
pred_dtr = dtr.predict(X_reg_test)
print(metrics.r2_score(y_reg_test, pred_dtr))
print(metrics.mean_squared_error(y_reg_test, pred_dtr))
print(metrics.mean_squared_error(y_reg_test, pred_dtr, squared=False))
print(metrics.mean_absolute_error(y_reg_test, pred_dtr))

## Обучение композиций

### Бэггинг

%%time
bagging_reg = ensemble.BaggingRegressor(dtr)
bagging_reg.fit(X_reg_train, y_reg_train)
pred_bagging_reg = bagging_reg.predict(X_reg_test)
print(metrics.r2_score(y_reg_test, pred_bagging_reg))
print(metrics.mean_squared_error(y_reg_test, pred_bagging_reg))
print(metrics.mean_squared_error(y_reg_test, pred_bagging_reg, squared=False))
print(metrics.mean_absolute_error(y_reg_test, pred_bagging_reg))

После применения бэггинга со стандартными параметрами точность сильно увеличилась.

bagging_reg_parameters_grid = { 'n_estimators' : [5, 10, 50, 100],
                               'max_samples' : range(1, 5, 1),
                               'max_features' : range(1, 5, 1)
}

bagging_reg_grid = GridSearchCV(bagging_reg, bagging_reg_parameters_grid,
                        cv=5, n_jobs=-1, verbose=True)
bagging_reg_grid.fit(X_reg_train, y_reg_train)
bagging_reg_grid.best_params_, bagging_reg_grid.best_score_

После подбора параметров точность очень сильно упала.

### Бустинг

boosting_reg = ensemble.GradientBoostingRegressor()
boosting_reg.fit(X_reg_train, y_reg_train)
pred_boosting_reg = boosting_reg.predict(X_reg_test)
print(metrics.r2_score(y_reg_test, pred_bagging_reg))
print(metrics.mean_squared_error(y_reg_test, pred_boosting_reg))
print(metrics.mean_squared_error(y_reg_test, pred_boosting_reg, squared=False))
print(metrics.mean_absolute_error(y_reg_test, pred_boosting_reg))

Бустинг без параметров тоже сильно поднял точность.

boosting_reg_parameters_grid = { 'loss' : ['ls', 'lad', 'huber', 'quantile'],
                                'learning_rate' : np.arange(0.1, 0.8, 0.1),
                                'n_estimators' : [10, 50, 100 , 500]
}

boosting_reg_grid = GridSearchCV(boosting_reg, boosting_reg_parameters_grid,
                        cv=5, n_jobs=-1, verbose=True)
boosting_reg_grid.fit(X_reg_train, y_reg_train)
boosting_reg_grid.best_params_, boosting_reg_grid.best_score_

После подбора качество только возросло.

### Стекинг

estimators_reg = [
  ('dtr', DecisionTreeRegressor()),
  ('ridge', Ridge())
]

stacking_reg = ensemble.StackingRegressor(estimators=estimators_reg)
stacking_reg.fit(X_reg_train, y_reg_train)
pred_stacking_reg = stacking_reg.predict(X_reg_test)
print(metrics.r2_score(y_reg_test, pred_stacking_reg))
print(metrics.mean_squared_error(y_reg_test, pred_stacking_reg))
print(metrics.mean_squared_error(y_reg_test, pred_stacking_reg, squared=False))
print(metrics.mean_absolute_error(y_reg_test, pred_stacking_reg))

stacking_reg_parameters_grid = { 'cv' : [1, 3, 5, 7, 9]
}

stacking_reg_grid = GridSearchCV(stacking_reg, stacking_reg_parameters_grid,
                        cv=5, n_jobs=-1, verbose=True)
stacking_reg_grid.fit(X_reg_train, y_reg_train)
stacking_reg_grid.best_params_, stacking_reg_grid.best_score_

Стекинг показывает неплохой результат но отстаёт от остальных композиций.

Лучшей композицией оказался бустинг после подбора параметров.
